requirements.txt
numpy==1.26.0
Flask==3.0.1
python-dotenv==1.0.1
langchain=0.2.5
langchain_openai=0.1.9
openai=1.35.3
libsql-experimental=0.0.34
langgraph=0.1.1